added gift and inspection apps/model/view/urls
