+++ gift/templates/award_card.html
@@ -1,0 +1,121 @@
+{% extends "base.html" %}
+{% load static %}
+{% block body %}
+{% include "navbar.html" %}
+
+    <!-- Flexbox to center the form -->
+        <div class="bg-white p-6 rounded-lg shadow-lg w-full max-w-md m-auto mt-6">
+            <h2 class="text-2xl font-bold text-gray-800 mb-4 text-center">Award Gift Card</h2>
+            
+            <form method="POST" class="space-y-4">
+                {% csrf_token %}
+                
+                <!-- Date Awarded Field -->
+                <div>
+                    <label for="{{ form.date_award.id_for_label }}" class="block text-sm font-medium text-gray-700">Date Awarded</label>
+                    <div class="mt-1">
+                        {{ form.date_award }}
+                    </div>
+                    {% if form.date_award.errors %}
+                        <div class="text-red-500 text-sm mt-2">
+                            {% for error in form.date_award.errors %}
+                                <p>{{ error }}</p>
+                            {% endfor %}
+                        </div>
+                    {% endif %}
+                </div>
+
+                 <!-- Employee Name Select Field -->
+                 <div>
+                    <label for="{{ form.employee_name.id_for_label }}" class="block text-sm font-medium text-gray-700">Employee Name</label>
+                    <div class="mt-1">
+                        {{ form.employee_name }}
+                    </div>
+                    {% if form.employee_name.errors %}
+                        <div class="text-red-500 text-sm mt-2">
+                            {% for error in form.employee_name.errors %}
+                                <p>{{ error }}</p>
+                            {% endfor %}
+                        </div>
+                    {% endif %}
+                </div>
+
+                <!-- Company Name Select Field -->
+                <div>
+                    <label for="{{ form.company_name.id_for_label }}" class="block text-sm font-medium text-gray-700">Company Name</label>
+                    <div class="mt-1">
+                        {{ form.company_name }}
+                    </div>
+                    {% if form.company_name.errors %}
+                        <div class="text-red-500 text-sm mt-2">
+                            {% for error in form.company_name.errors %}
+                                <p>{{ error }}</p>
+                            {% endfor %}
+                        </div>
+                    {% endif %}
+                </div>
+
+               
+                <!-- Amount Field -->
+                <div>
+                    <label for="{{ form.amount.id_for_label }}" class="block text-sm font-medium text-gray-700">Amount</label>
+                    <div class="mt-1">
+                        {{ form.amount }}
+                    </div>
+                    {% if form.amount.errors %}
+                        <div class="text-red-500 text-sm mt-2">
+                            {% for error in form.amount.errors %}
+                                <p>{{ error }}</p>
+                            {% endfor %}
+                        </div>
+                    {% endif %}
+                </div>
+
+                <!-- Card Number Field -->
+                 
+                <div>
+                    <label for="{{ form.card.id_for_label }}" class="block text-sm font-medium text-gray-700">Card Number</label>
+                    <div class="mt-1">
+                        {{ form.card }}
+                    </div>
+                    {% if form.card.errors %}
+                        <div class="text-red-500 text-sm mt-2">
+                            {% for error in form.card.errors %}
+                                <p>{{ error }}</p>
+                            {% endfor %}
+                        </div>
+                    {% endif %}
+                </div>
+                
+
+                <!-- Reason Field -->
+                <div>
+                    <label for="{{ form.reason.id_for_label }}" class="block text-sm font-medium text-gray-700">Reason for Gift Given</label>
+                    <div class="mt-1">
+                        {{ form.reason }}
+                    </div>
+                    {% if form.reason.errors %}
+                        <div class="text-red-500 text-sm mt-2">
+                            {% for error in form.reason.errors %}
+                                <p>{{ error }}</p>
+                            {% endfor %}
+                        </div>
+                    {% endif %}
+                </div>
+
+                <!-- Submit Button -->
+                <div class=" text-center justify-center">
+                    <button type="submit" class="bg-red-500 text-white px-6 py-2 rounded-lg hover:bg-red-600 focus:outline-none">
+                        Award Gift Card
+                    </button>
+                </div>
+            </form>
+
+            <div class="flex justify-between items-center mt-2 w-full">
+                <a href="{% url 'gift_card' %}" class="w-full py-2 bg-red-500 text-white rounded-lg hover:bg-red-600 focus:outline-none text-center block">
+                    Go to Gift Card
+                </a>
+            </div>
+            
+        </div>
+{% endblock %}
