+++ gift/templates/gift_card.html
@@ -1,0 +1,73 @@
+{% extends "base.html" %}
+{% load static %}
+{% block body %}
+{% include "navbar.html" %}
+
+    <div class="bg-white p-6 rounded-lg shadow-lg w-full max-w-md m-auto mt-12">
+        <h2 class="text-2xl font-bold text-gray-800 mb-4 text-center">Add Gift Card</h2>
+        
+        <form method="POST" class="space-y-4">
+            {% csrf_token %}
+
+           
+             <!-- Company Select Field -->
+             <div>
+                <label for="{{ form.company.id_for_label }}" class="block text-sm font-medium text-gray-700">Select Company</label>
+                <div class="mt-1">
+                    {{ form.company }}
+                </div>
+                {% if form.company.errors %}
+                    <div class="text-red-500 text-sm mt-2">
+                        {% for error in form.company.errors %}
+                            <p>{{ error }}</p>
+                        {% endfor %}
+                    </div>
+                {% endif %}
+            </div>
+
+            <!-- Date of Purchase Field -->
+            <div>
+                <label for="{{ form.date_of_purchase.id_for_label }}" class="block text-sm font-medium text-gray-700">Date of Purchase</label>
+                <div class="mt-1">
+                    {{ form.date_of_purchase }}
+                </div>
+                {% if form.date_of_purchase.errors %}
+                    <div class="text-red-500 text-sm mt-2">
+                        {% for error in form.date_of_purchase.errors %}
+                            <p>{{ error }}</p>
+                        {% endfor %}
+                    </div>
+                {% endif %}
+            </div>
+
+            <!-- Amount Field -->
+            <div>
+                <label for="{{ form.amount.id_for_label }}" class="block text-sm font-medium text-gray-700">Amount</label>
+                <div class="mt-1">
+                    {{ form.amount }}
+                </div>
+                {% if form.amount.errors %}
+                    <div class="text-red-500 text-sm mt-2">
+                        {% for error in form.amount.errors %}
+                            <p>{{ error }}</p>
+                        {% endfor %}
+                    </div>
+                {% endif %}
+            </div>
+
+            <!-- Submit Button -->
+            <div class="justify-center text-center">
+                <button type="submit" class="bg-red-500 text-white px-6 py-2 rounded-lg hover:bg-red-600 focus:outline-none ">
+                    Create Gift Card
+                </button>
+            </div>
+
+        </form>
+            <div class="flex justify-between items-center mt-2 w-full">
+                <a href="{% url 'award_card' %}" class="w-full py-2 bg-red-500 text-white rounded-lg hover:bg-red-600 focus:outline-none text-center block">
+                    Go to Award Card
+                </a>
+            </div>
+    </div>
+
+{% endblock %}
