+++ gift/views.py
@@ -1,0 +1,72 @@
+from django.shortcuts import render,redirect
+from django.views import View
+from rest_framework.permissions import IsAuthenticated
+from django.http import HttpResponseForbidden
+from authentication.models import UserProfile
+from gift.forms import AwardCardForm, GiftCardForm
+from datetime import datetime
+
+
+class GiftCardView(View):
+    permission_classes = [IsAuthenticated]
+
+    def dispatch(self, request, *args, **kwargs):
+        # Manually check permissions before executing the view logic
+        for permission in self.permission_classes:
+            permission_instance = permission()
+            if not permission_instance.has_permission(request, self):
+                return HttpResponseForbidden("You do not have permission to view this page.")
+        return super().dispatch(request, *args, **kwargs)
+
+    def get(self,request):
+        form = GiftCardForm()
+        return render(request, "gift_card.html",{"form":form})
+
+
+    def post(self,request):
+        form = GiftCardForm(request.POST)
+        if form.is_valid():
+            gift_card=form.save(commit=False)
+            current_user=request.user
+            user=UserProfile.objects.get(user=current_user)
+            gift_card.added_by=user
+            gift_card.save()
+            
+            return redirect('gift_card')
+        else:
+            form = GiftCardForm()
+        
+        return render(request, "gift_card.html",{"form":form})
+
+
+class AwardCardView(View):
+    permission_classes = [IsAuthenticated]
+
+    def dispatch(self, request, *args, **kwargs):
+        # Manually check permissions before executing the view logic
+        for permission in self.permission_classes:
+            permission_instance = permission()
+            if not permission_instance.has_permission(request, self):
+                return HttpResponseForbidden("You do not have permission to view this page.")
+        return super().dispatch(request, *args, **kwargs)
+
+    def get(self,request):
+        form = AwardCardForm()
+        return render(request, "award_card.html",{"form":form})
+
+
+    def post(self,request):
+        form = AwardCardForm(request.POST)
+        if form.is_valid():
+            gift_card=form.save(commit=False)
+            current_user=request.user
+            user=UserProfile.objects.get(user=current_user)
+            gift_card.awarded_by=user
+            gift_card.date_saved=datetime.now()
+            gift_card.save()
+            
+            return redirect('award_card')
+        else:
+            form = AwardCardForm()
+        
+        return render(request, "award_card.html",{"form":form})
