+++ inspection/templates/truck_inspection.html
