+++ station/forms.py
@@ -1,6 +1,7 @@
 from django import forms
 from vehicle.models import Vehicle
-from .models import Fleet_order, Trailer_inspection, Truck_inspection, Vehicle_inspection,Station_inspection
+from .models import Fleet_order,Vehicle_inspection,Station_inspection
+
 
 class StationInspectionForm(forms.ModelForm):
     class Meta:
@@ -129,205 +130,3 @@ class FleetOrderForm(forms.ModelForm):
             }),
         }
 
-
-class TrailerInspectionForm(forms.ModelForm):
-    class Meta:
-        model = Trailer_inspection
-        fields = [
-            'date', 'trailer', 'clean_status', 'trash', 
-            'blanket_84', 'hand_trucks_with_covers', 'four_wheel_dolly', 
-            'short_straps', 'long_straps', 'ramp', 'rubber_bands', 
-            'red_floor_runner', 'forearm_straps', 'wardrobe_boxes_with_bars', 
-            'tv_box_for_rental', 'multi_tool_set', 'hand_tools_bag', 
-            'two_carabiner', 'broom'
-        ]
-        widgets = {
-            'date': forms.DateInput(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-                'type': 'date',
-            }),
-            'trailer': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'clean_status': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'trash': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'blanket_84': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'hand_trucks_with_covers': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'four_wheel_dolly': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'short_straps': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'long_straps': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'ramp': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'rubber_bands': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'red_floor_runner': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'forearm_straps': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'wardrobe_boxes_with_bars': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'tv_box_for_rental': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'multi_tool_set': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'hand_tools_bag': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'two_carabiner': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'broom': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'saved_by': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            })
-        }
-
-
-class TruckInspectionForm(forms.ModelForm):
-    class Meta:
-        model = Truck_inspection
-        fields = [
-            'date', 'truck', 'clean_status', 'in_cab', 'bed_of_truck', 
-            'first_aid_kit', 'floor_mats', 'business_cards', 'business_cards_magnetic', 
-            'fuses', 'two_pens', 'sharpie', 'camera', 'flash_light', 'sun_visor', 
-            'geo_tab', 'jack_and_links', 'cab_card', 'registration', 'insurance_card', 
-            'accident_report_form', 'process_of_accident', 'fire_extinguisher', 
-            'expiry_date_fe', 'four_way', 'min_7_orange', 'hazard_triangle_x3', 
-            'jumper_cables', 'large_door_stops', 'trash_bag', 'roll_paper_towels', 
-            'small_hand', 'bottle_jack', 'cones', 'spare_tire', 'condition_spare_tyre'
-        ]
-        widgets = {
-            'date': forms.DateInput(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-                'type': 'date',
-            }),
-            'truck': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'clean_status': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'in_cab': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'bed_of_truck': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'first_aid_kit': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'floor_mats': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'business_cards': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'business_cards_magnetic': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'fuses': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'two_pens': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'sharpie': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'camera': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'flash_light': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'sun_visor': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'geo_tab': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'jack_and_links': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'cab_card': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'registration': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'insurance_card': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'accident_report_form': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'process_of_accident': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'fire_extinguisher': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'expiry_date_fe': forms.DateInput(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-                'type': 'date',
-            }),
-            'four_way': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'min_7_orange': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'hazard_triangle_x3': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'jumper_cables': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'large_door_stops': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'trash_bag': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'roll_paper_towels': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'small_hand': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'bottle_jack': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'cones': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'spare_tire': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            }),
-            'condition_spare_tyre': forms.Select(attrs={
-                'class': 'border border-gray-300 rounded-md px-2 py-2 focus:outline-none focus:ring-2 focus:ring-red-500 w-full',
-            })
-        }
