+++ station/models.py
@@ -3,6 +3,7 @@ from authentication.models import UserProfile
 from vehicle.models import Vehicle
 
 
+
 class Station(models.Model):
     name = models.CharField(max_length=100)
     location = models.CharField(max_length=255, blank=True, null=True)
@@ -160,132 +161,3 @@ class Fleet_order(models.Model):
 
     def __str__(self):
         return f"{self.urgency_level.capitalize()} Order - {self.status} on {self.date}"
-
-
-class Truck_inspection(models.Model):
-
-    CLEAN_STATUS_CHOICES = [
-        ('full of trash & dirt', 'Full of Trash & Dirt'),
-        ('free of trash', 'Free of Trash'),
-        ('removed trash', 'Removed Trash'),
-    ]
-    TRASH_CLEAN_STATUS_CHOICES=[
-        ('not clean','Not Clean'),
-        ('dirty now clean','Dirty Now Clean'),
-        ('already clean','Already Clean'),
-    ]
-    GENERAL_CHOICES = [
-        ('missing-not restocked', 'Missing-Not Restocked'),
-        ('missing restocked', 'Missing Restocked'),
-        ('present', 'Present'),
-    ]
-    SPARE_TYRE_CHOICES=[
-        ('good','Good'),
-        ('normal','Normal'),
-        ('damage','Damage'),
-    ]
-
-    # Truck and Date Information
-    date = models.DateField(null=True, blank=True)
-    truck = models.ForeignKey(Vehicle, on_delete=models.CASCADE,null=True, blank=True,limit_choices_to={'vehicle_type': 'truck'})
-    
-    # Trash & Clean Status
-    clean_status = models.CharField(max_length=50, choices=CLEAN_STATUS_CHOICES, null=True, blank=True)
-    in_cab = models.CharField(max_length=50, choices=TRASH_CLEAN_STATUS_CHOICES, null=True, blank=True)
-    bed_of_truck = models.CharField(max_length=50, choices=TRASH_CLEAN_STATUS_CHOICES, null=True, blank=True)
-    
-    # In Cab
-    first_aid_kit = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    floor_mats = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    business_cards = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    business_cards_magnetic = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    fuses = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    two_pens = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    sharpie = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    camera = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    flash_light = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    sun_visor = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    geo_tab = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    jack_and_links = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    cab_card = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    registration = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    insurance_card = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    accident_report_form = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    process_of_accident = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-
-    # Tool Box
-    fire_extinguisher = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    expiry_date_fe = models.DateField(null=True, blank=True)
-    four_way = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    min_7_orange = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    hazard_triangle_x3 = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    jumper_cables = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    large_door_stops = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    trash_bag = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    roll_paper_towels = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    small_hand = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    bottle_jack = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    
-    # Equipment Checklist
-    cones = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    spare_tire = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    condition_spare_tyre = models.CharField(max_length=50, choices=SPARE_TYRE_CHOICES, null=True, blank=True)
-
-    # saving information
-    saved_by = models.ForeignKey(UserProfile, on_delete=models.CASCADE, null=True, blank=True)
-    date_saved = models.DateTimeField(auto_now_add=True)
-
-    def __str__(self):
-        return f"Truck {self.truck} Inspection on {self.date}"
-
-
-class Trailer_inspection(models.Model):
-
-    CLEAN_STATUS_CHOICES = [
-        ('full of trash & dirt', 'Full of Trash & Dirt'),
-        ('free of trash', 'Free of Trash'),
-        ('removed trash', 'Removed Trash'),
-        ('detailed inside/out', 'Detailed Inside/Out'),
-        ('wiped down inside/ washed outside', 'Wiped Down Inside/ Washed Outside'),
-        ('wiped down inside only', 'Wiped Down Inside Only'),
-        ('washed outside only', 'Washed Outside Only'),
-    ]
-    GENERAL_CHOICES = [
-        ('missing-not restocked', 'Missing-Not Restocked'),
-        ('missing restocked', 'Missing Restocked'),
-        ('present', 'Present'),
-    ]
-    
-    # Trailer and Date Information
-    date = models.DateField(null=True, blank=True)
-    trailer = models.ForeignKey(Vehicle, on_delete=models.CASCADE,null=True, blank=True,limit_choices_to={'vehicle_type': 'trailer'})
-    
-    # Trailer Trash Status
-    clean_status = models.CharField(max_length=50, choices=CLEAN_STATUS_CHOICES, null=True, blank=True)
-    trash = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    
-    # Equipment Checklist
-    blanket_84 = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    hand_trucks_with_covers = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    four_wheel_dolly = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    short_straps = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    long_straps = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    ramp = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    rubber_bands = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    red_floor_runner = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    forearm_straps = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    wardrobe_boxes_with_bars = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    tv_box_for_rental = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    multi_tool_set = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    hand_tools_bag = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    two_carabiner = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-    broom = models.CharField(max_length=50, choices=GENERAL_CHOICES, null=True, blank=True)
-
-    # Saving Information
-    saved_by = models.ForeignKey(UserProfile, on_delete=models.CASCADE, null=True, blank=True) 
-    date_saved = models.DateTimeField(auto_now_add=True)
-    
-    def __str__(self):
-        return f"Trailer {self.trailer} Inspection on {self.date}"
-
-
