+++ station/templates/station_base.html
@@ -1,12 +1,13 @@
+{% extends "base.html" %}
 {% load static %}
+{% block body %}
 {% include "navbar.html" %}
 
-<!DOCTYPE html>
-<html lang="en">
+
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Fire House Sidebar</title>
+    <title>Fire House</title>
     <link rel="icon" href="{% static 'images/fire_house_logo.svg' %}" type="image/icon type">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0-beta3/css/all.min.css">
     <script src="https://cdn.tailwindcss.com"></script>
@@ -212,3 +213,4 @@
     </div>
 </body>
 </html>
+{% endblock %}
