+++ vehicle/templates/availability_logistic_base.html
@@ -1,18 +1,7 @@
+{% extends "base.html" %}
 {% load static %}
+{% block body %}
 {% include "navbar.html" %}
-
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Fire House</title>
-    <link rel="icon" href="{% static 'images/fire_house_logo.svg' %}" type="image/icon type">
-    <script src="https://cdn.tailwindcss.com"></script>
-    <link href="https://fonts.googleapis.com/css2?family=YourCustomFont:wght@400;700&display=swap" rel="stylesheet">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0-beta3/css/all.min.css">
-</head>
-<body style="font-family: var(--titleFont);">
 
     <div class="flex h-screen">
         <!-- Sidebar -->
@@ -61,5 +50,5 @@
             {% endblock %}
         </main>
     </div>
-</body>
-</html>
+
+{% endblock %}
